feat(backend): EmailService, EstadoTerminal, tests unitarios y ajustes Bloque 7
- Agrega EmailService (@Async, Gmail SMTP) para recuperacion de password temporal
- Agrega EstadoTerminal enum y TerminalEstadoResponse DTO
- Ajustes en AuthService: generacion de password temporal de 8 caracteres
- Ajustes en services: InformeService, PdfService, PresenciaService, SaldoService, TerminalService
- Ajustes en controllers: EmpleadoController, TerminalController
- Ajustes en entidad Empleado, DTOs request/response y GlobalExceptionHandler
- Actualiza application-dev.yml, application-mysql.yml y data.sql
- Agrega 34 tests unitarios: JwtTokenProviderTest, SaldoServiceTest, TerminalServiceTest, GlobalExceptionHandlerTest

+++ staffflow-backend/src/main/java/com/staffflow/domain/entity/Empleado.java
@@ -56,8 +56,9 @@ public class Empleado {
     private CategoriaEmpleado categoria;
 
     // Horas semanales contractuales. Base para el cálculo de saldo de horas.
+    // Double para soportar jornadas parciales con decimales (ej: 37.5h).
     @Column(name = "jornada_semanal_horas", nullable = false)
-    private Integer jornadaSemanalHoras = 40;
+    private Double jornadaSemanalHoras = 40.0;
 
     // Minutos de jornada diaria esperada. Usado en el cálculo de jornada efectiva.
     @Column(name = "jornada_diaria_minutos", nullable = false)
